fix: PR comment resolved
