+++ packages/components/FilePreview/AssetsPagination/PaginationProps.type.ts
@@ -1,0 +1,5 @@
+export interface PaginationProps {
+  currentPage: number;
+  maxPage: number;
+  onChangePage: (page: number) => void;
+}
