+++ packages/components/FilePreview/AssetsPagination/RightContainer.tsx
@@ -1,102 +1,19 @@
 import React from "react";
-import { TextInput, TouchableOpacity, View } from "react-native";
+import { TouchableOpacity, View } from "react-native";
 
-import { BrandText } from "./BrandText";
-import { SVG } from "./SVG";
-import { SEARCH_BAR_INPUT_HEIGHT } from "./Search/SearchBarInput";
-import { TertiaryBox } from "./boxes/TertiaryBox";
-import { SpacerRow } from "./spacer";
-import chevronLeftDoubleSVG from "../../assets/icons/chevron-left-double.svg";
-import chevronLeftSVG from "../../assets/icons/chevron-left.svg";
-import chevronRightDoubleSVG from "../../assets/icons/chevron-right-double.svg";
-import chevronRightSVG from "../../assets/icons/chevron-right.svg";
-import { neutral17, neutral77, primaryColor } from "../utils/style/colors";
-import { fontSemibold14 } from "../utils/style/fonts";
-import { layout } from "../utils/style/layout";
+import { PaginationProps } from "./PaginationProps.type";
+import chevronLeftDoubleSVG from "../../../../assets/icons/chevron-left-double.svg";
+import chevronLeftSVG from "../../../../assets/icons/chevron-left.svg";
+import chevronRightDoubleSVG from "../../../../assets/icons/chevron-right-double.svg";
+import chevronRightSVG from "../../../../assets/icons/chevron-right.svg";
+import { primaryColor } from "../../../utils/style/colors";
+import { fontSemibold14 } from "../../../utils/style/fonts";
+import { BrandText } from "../../BrandText";
+import { SVG } from "../../SVG";
+import { TertiaryBox } from "../../boxes/TertiaryBox";
+import { SpacerRow } from "../../spacer";
 
-interface PaginationProps {
-  currentPage: number;
-  maxPage: number;
-  onChangePage: (page: number) => void;
-}
-
-const LeftContainer = ({
-  currentPage,
-  maxPage,
-  onChangePage,
-}: PaginationProps) => {
-  return (
-    <View style={{ flex: 1, flexDirection: "row" }}>
-      <View
-        style={{
-          flexDirection: "row",
-          alignItems: "center",
-          justifyContent: "flex-start",
-        }}
-      >
-        <BrandText
-          style={{
-            ...fontSemibold14,
-            color: neutral77,
-            paddingRight: layout.spacing_x1,
-            lineHeight: 14,
-          }}
-        >
-          Page {currentPage + 1} of {maxPage}
-        </BrandText>
-      </View>
-      <View
-        style={{
-          flexDirection: "row",
-          alignItems: "center",
-          justifyContent: "flex-start",
-        }}
-      >
-        <BrandText
-          style={{
-            ...fontSemibold14,
-            color: neutral77,
-            paddingRight: layout.spacing_x1,
-            lineHeight: 14,
-          }}
-        >
-          | Go to page:
-        </BrandText>
-        <TertiaryBox
-          style={[
-            {
-              flexDirection: "row",
-              paddingHorizontal: layout.spacing_x1_5,
-              backgroundColor: neutral17,
-              width: 80,
-              height: SEARCH_BAR_INPUT_HEIGHT,
-              alignItems: "center",
-            },
-          ]}
-        >
-          <TextInput
-            defaultValue={(currentPage + 1).toString()}
-            inputMode="numeric"
-            style={[
-              fontSemibold14,
-              {
-                color: "white",
-                flex: 1,
-                width: 56,
-              },
-              { outlineStyle: "none" } as any,
-            ]}
-            onSubmitEditing={(value) => {
-              onChangePage(+value.nativeEvent.text - 1);
-            }}
-          />
-        </TertiaryBox>
-      </View>
-    </View>
-  );
-};
-
-const RightContainer = ({
+export const RightContainer = ({
   currentPage,
   maxPage,
   onChangePage,
@@ -236,48 +153,6 @@ const RightContainer = ({
           </TertiaryBox>
         </TouchableOpacity>
       </View>
-    </View>
-  );
-};
-
-export const AssetsPagination = ({
-  currentPage,
-  maxPage,
-  onChangePage,
-}: PaginationProps) => {
-  const handleChangePage = (pageIndex: number) => {
-    if (pageIndex < 0) {
-      pageIndex = 0;
-    } else if (pageIndex >= maxPage) {
-      pageIndex = maxPage - 1;
-    }
-    if (pageIndex !== currentPage) {
-      onChangePage(pageIndex);
-    }
-  };
-
-  return (
-    <View
-      style={{
-        flexDirection: "row",
-        justifyContent: "space-around",
-        width: "100%",
-        paddingHorizontal: layout.spacing_x2,
-      }}
-    >
-      <LeftContainer
-        currentPage={currentPage}
-        maxPage={maxPage}
-        onChangePage={handleChangePage}
-      />
-
-      <SpacerRow size={1} />
-
-      <RightContainer
-        currentPage={currentPage}
-        maxPage={maxPage}
-        onChangePage={handleChangePage}
-      />
     </View>
   );
 };
